+++ wattsquad/ml_logic/calculations.py
@@ -367,52 +367,6 @@ def cost_savings(flexibility_degree):
 #my_cost_savings = cost_savings(0.9)
 #print(my_cost_savings)
 
-def actual_battery_percentage(capacity=500, initial_battery_percentage=0):
-    """
-    Calculate the actual battery percentage based on production and consumption.
-
-    Args:
-        capacity (float): Battery capacity in kWh (default is 500 kWh).
-        initial_battery_percentage (float): Initial battery percentage (default is 0).
-
-    Returns:
-        DataFrame: Updated DataFrame with columns for battery status and grid purchases.
-    """
-    # Load data
-    data = load_training_data()
-
-    # Initialize new columns
-    data['excess_production_kwH'] = data['actual_production'] - data['actual_consumption']
-    data['battery_percentage'] = initial_battery_percentage
-    data['electricity_bought_kwH'] = 0
-    data['electricity_bought_NOK'] = 0
-
-    # Iterate over each time step
-    for i in range(len(data)):
-        excess_production = data.loc[i, 'excess_production_kwH']
-        battery_percentage = data.loc[i, 'battery_percentage']
-
-        if excess_production > 0: # Excess production, charge the battery
-            additional_charge = excess_production / capacity
-            new_battery_percentage = battery_percentage + additional_charge
-            data.loc[i, 'battery_percentage'] = min(new_battery_percentage, 1) # Cap at 100%
-
-        elif excess_production < 0: # Deficit, discharge battery or buy from the grid
-            required_discharge = abs(excess_production) /  capacity
-            if battery_percentage >= required_discharge: # Sufficient battery
-                data.loc[i, 'battery_percentage'] = battery_percentage - required_discharge
-            else: # Insufficient battery, buy from grid
-                data.loc[i, 'battery_percentage'] = 0
-                shortfall_kwH = abs(excess_production) - (battery_percentage * capacity)
-                data.loc[i, 'electricity_bought_kwH'] = shortfall_kwH
-                data.loc[i, 'electricity_bought_NOK'] = shortfall_kwH * data.loc[i, 'electricity_price']
-
-        # Carry forward the battery percentage to the next row
-        if i < len(data) -1:
-            data.loc[i+1, 'battery_percentage'] = data.loc[i, 'battery_percentage']
-
-    return data
-
 def pred_battery_percentage(capacity=500, initial_battery_percentage=0):
     """
     Calculate the predicted battery percentage based on production and consumption.
@@ -466,7 +420,7 @@ def pred_battery_percentage(capacity=500, initial_battery_percentage=0):
 
 
 
-def cost_saving(flexibility_degree):
+def cost_saving(flexibility_degree: float):
     # 1. Load data
     data = load_data()
